add 'auth' directive; create directives folder
